TST: Consolidate groupby any, all tests (#40811)
